% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector based on specified minimum and maximum values}
\usage{
normalize(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
A numeric vector of the same length as \code{x}
}
\description{
This function normalizes a vector based on specified minimum
             and maximum values. If the minimum and maximum values are not
             specified, the function will use the minimum and maximum values
             of the vector (ignoring missing values).
}
\examples{
normalize(c(0, 5, 10))
normalize(c(1, 1, 1))
normalize(c(NA, 2, 3))
}
\author{
E. F. Haghish
}
