% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{owen}
\alias{owen}
\title{Compute Owen Value Decompositions}
\usage{
owen(vfun, factors, outcomes = "value", silent = FALSE, ...)
}
\arguments{
\item{vfun}{A value function.}

\item{factors}{A list of vectors, where each item of the list is treated
as one group. This list also can contain further lists, defining subgroups. Only two levels
of nesting are currently supported.}

\item{outcomes}{Column names for outcome values (usually this is only one).}

\item{silent}{If FALSE (the default), prints a progress bar.}

\item{...}{Additional arguments passed to \code{vfun}.}
}
\description{
Compute Owen Value Decompositions
}
