% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley}
\alias{shapley}
\title{Compute Shapley-Shorrocks Value Decompositions}
\usage{
shapley(vfun, factors, silent = FALSE)
}
\arguments{
\item{vfun}{A value function.}

\item{factors}{A vector of factors, passed to \code{vfun}.}

\item{silent}{If FALSE (the default), prints a progress bar.}
}
\value{
Returns a data frame with N rows, where N is the number of factors.
}
\description{
Compute Shapley-Shorrocks Value Decompositions
}
\references{
Shapley, L. S. (1953). A value for n-person games.
     Contributions to the Theory of Games, 2(28), 307-317.

Shorrocks, A. F. (2013).
     Decomposition procedures for distributional analysis:
     a unified framework based on the Shapley value. Journal of Economic Inequality, 1-28.
}
