% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley_sampled}
\alias{shapley_sampled}
\title{Compute Sampled Shapley/Owen Value Decompositions}
\usage{
shapley_sampled(
  vfun,
  factors,
  outcomes = "value",
  last_n = 100,
  precision = 1e-04,
  max_iter = 1e+06,
  silent = FALSE,
  ...
)
}
\arguments{
\item{vfun}{A value function.}

\item{factors}{A vector of factors, passed to \code{vfun}. List for Owen values is allowed,
but only one level.}

\item{outcomes}{Column names for outcome values (usually this is only one).}

\item{last_n}{An integer that specifies on how many values the standard deviation of the last
samples is calculated.}

\item{precision}{The stopping criterion.}

\item{max_iter}{Maximum number of samples to be drawn for each factor. Should be reasonably large
and usually doesn't matter. To adjust computation speed,
it is preferred to adjust \code{precision}.}

\item{silent}{If FALSE (the default), prints a progress bar.}

\item{...}{Additional arguments passed to \code{vfun}.}
}
\description{
Compute Sampled Shapley/Owen Value Decompositions
}
