% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Weighted Permutation Test for Difference of Means}
\usage{
test(var1, var2, weights, n = 2000)
}
\arguments{
\item{var1}{A numeric vector.}

\item{var2}{A numeric vector of the same length as \code{var1}.}

\item{weights}{A numeric vector of weights, assumed to be the same for both \code{var1} and \code{var2}.}

\item{n}{The number of permutations to perform (default is 1000).}
}
\value{
A list containing the observed difference in means and the p-value of the test.
}
\description{
This function performs a weighted permutation test to determine if there is a significant
difference between the means of two weighted numeric vectors. It tests the null hypothesis
that the difference in means is zero against the alternative that it is not zero.
}
