% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector based on specified minimum and maximum values}
\usage{
normalize(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
normalized numeric vector
}
\description{
This function normalizes a vector based on specified minimum
             and maximum values. If the minimum and maximum values are not
             specified, the function will use the minimum and maximum values
             of the vector.
}
\author{
E. F. Haghish
}
