% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.feature.selection.R
\name{shapley.feature.selection}
\alias{shapley.feature.selection}
\title{Selects the top features with highest weighted mean shap values based on the
       specified criteria}
\usage{
shapley.feature.selection(
  shapley,
  method = "lowerCI",
  cutoff = 0,
  top_n_features = NULL
)
}
\arguments{
\item{shapley}{shapley object}

\item{cutoff}{numeric, specifying the cutoff for the method used for selecting
the top features. the default is zero, which means that all
features with the "method" criteria above zero will be selected.}

\item{top_n_features}{integer. if specified, the top n features with the
highest weighted SHAP values will be selected, overrullung
the 'cutoff' and 'method' arguments.}
}
\value{
normalized numeric vector
}
\description{
This function specifies the top features and prepares the data
             for plotting SHAP contributions for each row, or summary of absolute
             SHAP contributions for each feature.
}
\author{
E. F. Haghish
}
