% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley_sampled}
\alias{shapley_sampled}
\title{EXPERIMENTAL: Compute Sampled Shapley-Shorrocks Value Decompositions}
\usage{
shapley_sampled(vfun, factors, last_n = 100, precision = 1e-04,
  max_iter = 1e+06, silent = FALSE, ...)
}
\arguments{
\item{vfun}{A value function.}

\item{factors}{A vector of factors, passed to \code{vfun}.}

\item{last_n}{An integer that specifies on how many values the standard deviation of the last
samples is calculated.}

\item{precision}{The stopping criterion.}

\item{max_iter}{Maximum number of samples to be drawn for each factor. Should be reasonably large
and usually doesn't matter. To adjust computation speed,
it is preferred to adjust \code{precision}.}

\item{silent}{If FALSE (the default), prints a progress bar.}

\item{...}{Additional arguments passed to \code{vfun}.}
}
\value{
Returns a data frame with N rows, where N is the number of factors.
}
\description{
EXPERIMENTAL: Compute Sampled Shapley-Shorrocks Value Decompositions
}
\references{
Shapley, L. S. (1953). A value for n-person games.
     Contributions to the Theory of Games, 2(28), 307-317.

Shorrocks, A. F. (2013).
     Decomposition procedures for distributional analysis:
     a unified framework based on the Shapley value. Journal of Economic Inequality, 1-28.
}
